cleanup of book example, continue rules 3/4
to be continued
Book example is a clusterf--- (checking the book repo, it is only
complete enough to compile anyway) (of questionable use as an example)
(there must be a better one?). Rule 3: One level of abstraction per
Function, with 'BreakfastMaker' example. Rule 4: stepdown rule, begin
examination of 'grouping similar/top-level functions' issue, (and
additionally dump on topic).